+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Fine-grained selector permission.
 
 ## [0.2.10] - 2025-06-19
 ### Fixed
@@ -48,6 +50,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Using wrong fallback command node requirement
+
+### Changed
+- `minecraft.command.<command>.*` are granted by default
 
 ## [0.2.1] - 2023-07-01
 ### Fixed